ExpectaSupport: pass object by pointer instead of vararg.
The whole point of passing a _single_ object via `vararg` seems
redundant and risky, since there are default argument promotions in
effect which may change the input value. In particular, the address
sanitizer complained about passing a struct with three floats and
claimed that accessing the vararg causing out of memory access.

Since it's unclear whether calling `va_arg` on a struct is even
well-defined, I chose to drop vararg handling completely and use a
`void *` instead.

Additionally, I used `NSGetSizeAndAlignment` to make the floating point
struct handling more robust.

In general, I believe that most of the code in this method can be
reduced to a single `[NSValue valueWithBytes:objCType:]` call, but
that's out of the scope of this commit.