Expecta: pass actual to block matchers.
The current implementation of Expecta passes `actual` as a captured
variable to the `prerequisite`, `match`, `failureMessageForTo` and
`failureMessageForNotTo` blocks, and overwrites `actual` via a pointer
to an ObjC object.

From experimentation, it seems that matchers that are written in ARC do
not work well with this method - if `will` or `after` is used, `actual`
will not get updated on each iteration of the matcher. Additionally, we
suspect that this implementation may create memory leaks, as the object
pointed to is not being explicitly `release`ed.

This commit makes the code more straightforward and clean, by passing
`actual` to the aforementioned blocks, and removing it from the capture
list of these blocks.

Since this is an Expecta API change, it requires changes across all the
matchers.